backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfminer.six==20231228
docx2txt==0.8
spacy==3.7.2
sentence-transformers==2.2.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
